wip: split gui shell components
