+++ packages/gui-web/src/ScreenshotCaptureNotification.tsx
@@ -1,4 +1,4 @@
-import type { ReactElement } from "react";
+import { type ReactElement, useEffect, useState } from "react";
 
 export interface ScreenshotCapturedDetail {
   path: string;
@@ -30,4 +30,22 @@ export function ScreenshotCaptureNotification({ notification, onDismiss }: { not
       <button aria-label="Dismiss screenshot notification" onClick={onDismiss} type="button">×</button>
     </div>
   );
+}
+
+export function ScreenshotCaptureNotificationHost(): ReactElement | null {
+  const [screenshotNotification, setScreenshotNotification] = useState<ScreenshotCapturedDetail | undefined>();
+
+  useEffect(() => {
+    const showScreenshotNotification = (event: Event) => {
+      const detail = (event as CustomEvent<Partial<ScreenshotCapturedDetail>>).detail;
+      if (detail !== undefined && typeof detail.path === "string" && typeof detail.url === "string") {
+        setScreenshotNotification({ path: detail.path, url: detail.url });
+      }
+    };
+
+    window.addEventListener("aidevops:screenshot-captured", showScreenshotNotification);
+    return () => window.removeEventListener("aidevops:screenshot-captured", showScreenshotNotification);
+  }, []);
+
+  return screenshotNotification ? <ScreenshotCaptureNotification notification={screenshotNotification} onDismiss={() => setScreenshotNotification(undefined)} /> : null;
 }
